+++ .task/lib/index.js
@@ -1,5 +1,8 @@
 'use strict';
 
+var Q = require('q');
+var prompt = require('promptly').prompt;
+
 // Iterate over each task that is declared
 var forEachTask = exports.forEachTask = function(callback, command, tasks) {
   if(tasks == null && command == null) {
@@ -16,4 +19,19 @@ var forEachTask = exports.forEachTask = function(callback, command, tasks) {
     var subCommand = command ? command + ':' : '';
     forEachTask(callback, subCommand + index, tasks[index]);
   }
+};
+
+exports.prompt = function(message, opts) {
+  var deferred = Q.defer();
+
+  prompt(message, opts, function(err, name) {
+    if(err) {
+      deferred.reject(err);
+    }
+    else {
+      deferred.resolve(name);
+    }
+  });
+
+  return deferred.promise;
 };